removing non-used muon MVAs
