+++ DataFormats/PatCandidates/src/Muon.cc
@@ -31,8 +31,6 @@ Muon::Muon()
       pfEcalEnergy_(0),
       jetPtRatio_(0),
       jetPtRel_(0),
-      mvaValue_(0),
-      lowptMvaValue_(0),
       mvaIDValue_(0),
       softMvaValue_(0),
       inverseBeta_(0),
@@ -63,8 +61,6 @@ Muon::Muon(const reco::Muon& aMuon)
       pfEcalEnergy_(0),
       jetPtRatio_(0),
       jetPtRel_(0),
-      mvaValue_(0),
-      lowptMvaValue_(0),
       mvaIDValue_(0),
       softMvaValue_(0),
       inverseBeta_(0),
@@ -95,8 +91,6 @@ Muon::Muon(const edm::RefToBase<reco::Muon>& aMuonRef)
       pfEcalEnergy_(0),
       jetPtRatio_(0),
       jetPtRel_(0),
-      mvaValue_(0),
-      lowptMvaValue_(0),
       mvaIDValue_(0),
       softMvaValue_(0),
       inverseBeta_(0),
@@ -127,8 +121,6 @@ Muon::Muon(const edm::Ptr<reco::Muon>& aMuonRef)
       pfEcalEnergy_(0),
       jetPtRatio_(0),
       jetPtRel_(0),
-      mvaValue_(0),
-      lowptMvaValue_(0),
       mvaIDValue_(0),
       softMvaValue_(0),
       inverseBeta_(0),
