+++ PhysicsTools/PatAlgos/python/producersLayer1/muonProducer_cfi.py
@@ -109,13 +109,10 @@ patMuons = cms.EDProducer("PATMuonProducer",
     # Standard Muon Selectors and Jet-related observables
     # Depends on MiniIsolation, so only works in miniaod
     # Don't forget to set flags properly in miniAOD_tools.py                      
-    computeMuonMVA = cms.bool(False),
     computeMuonIDMVA = cms.bool(False),
-    mvaTrainingFile      = cms.FileInPath("RecoMuon/MuonIdentification/data/mu_2017_BDTG.weights.xml"),
     mvaIDTrainingFile      = cms.FileInPath("RecoMuon/MuonIdentification/data/mvaID.onnx"),
-    lowPtmvaTrainingFile = cms.FileInPath("RecoMuon/MuonIdentification/data/mu_lowpt_BDTG.weights.xml"),
     recomputeBasicSelectors = cms.bool(True),
-    mvaUseJec = cms.bool(True),
+    useJec = cms.bool(True),
     mvaDrMax = cms.double(0.4),
     mvaJetTag = cms.InputTag("pfCombinedInclusiveSecondaryVertexV2BJetTags"),
     mvaL1Corrector = cms.InputTag("ak4PFCHSL1FastjetCorrector"),
